Project2/deepmind-research/meshgraphnets/requirements.txt
tensorflow>=2.0
dm-sonnet<2
matplotlib
absl-py
numpy
